move type based functions into types.go
